+++ bpstudio-20210931/alibabacloud_bpstudio20210931/models.py
@@ -49,20 +49,20 @@ class CreateApplicationRequestInstances(TeaModel):
 class CreateApplicationRequest(TeaModel):
     def __init__(
         self,
-        application_params: Dict[str, Any] = None,
         area_id: str = None,
         client_token: str = None,
+        configuration: Dict[str, Any] = None,
         instances: List[CreateApplicationRequestInstances] = None,
         name: str = None,
         resource_group_id: str = None,
         template_id: str = None,
-        variables: str = None,
+        variables: Dict[str, Any] = None,
     ):
-        self.application_params = application_params
         # 区域ID
         self.area_id = area_id
         # 幂等标记
         self.client_token = client_token
+        self.configuration = configuration
         # 待替换实例列表
         self.instances = instances
         # 新建应用名
@@ -85,12 +85,12 @@ class CreateApplicationRequest(TeaModel):
             return _map
 
         result = dict()
-        if self.application_params is not None:
-            result['ApplicationParams'] = self.application_params
         if self.area_id is not None:
             result['AreaId'] = self.area_id
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+        if self.configuration is not None:
+            result['Configuration'] = self.configuration
         result['Instances'] = []
         if self.instances is not None:
             for k in self.instances:
@@ -107,12 +107,12 @@ class CreateApplicationRequest(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        if m.get('ApplicationParams') is not None:
-            self.application_params = m.get('ApplicationParams')
         if m.get('AreaId') is not None:
             self.area_id = m.get('AreaId')
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+        if m.get('Configuration') is not None:
+            self.configuration = m.get('Configuration')
         self.instances = []
         if m.get('Instances') is not None:
             for k in m.get('Instances'):
@@ -132,20 +132,20 @@ class CreateApplicationRequest(TeaModel):
 class CreateApplicationShrinkRequest(TeaModel):
     def __init__(
         self,
-        application_params_shrink: str = None,
         area_id: str = None,
         client_token: str = None,
+        configuration_shrink: str = None,
         instances_shrink: str = None,
         name: str = None,
         resource_group_id: str = None,
         template_id: str = None,
-        variables: str = None,
+        variables_shrink: str = None,
     ):
-        self.application_params_shrink = application_params_shrink
         # 区域ID
         self.area_id = area_id
         # 幂等标记
         self.client_token = client_token
+        self.configuration_shrink = configuration_shrink
         # 待替换实例列表
         self.instances_shrink = instances_shrink
         # 新建应用名
@@ -154,7 +154,7 @@ class CreateApplicationShrinkRequest(TeaModel):
         self.resource_group_id = resource_group_id
         # 模板ID
         self.template_id = template_id
-        self.variables = variables
+        self.variables_shrink = variables_shrink
 
     def validate(self):
         pass
@@ -165,12 +165,12 @@ class CreateApplicationShrinkRequest(TeaModel):
             return _map
 
         result = dict()
-        if self.application_params_shrink is not None:
-            result['ApplicationParams'] = self.application_params_shrink
         if self.area_id is not None:
             result['AreaId'] = self.area_id
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+        if self.configuration_shrink is not None:
+            result['Configuration'] = self.configuration_shrink
         if self.instances_shrink is not None:
             result['Instances'] = self.instances_shrink
         if self.name is not None:
@@ -179,18 +179,18 @@ class CreateApplicationShrinkRequest(TeaModel):
             result['ResourceGroupId'] = self.resource_group_id
         if self.template_id is not None:
             result['TemplateId'] = self.template_id
-        if self.variables is not None:
-            result['Variables'] = self.variables
+        if self.variables_shrink is not None:
+            result['Variables'] = self.variables_shrink
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        if m.get('ApplicationParams') is not None:
-            self.application_params_shrink = m.get('ApplicationParams')
         if m.get('AreaId') is not None:
             self.area_id = m.get('AreaId')
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+        if m.get('Configuration') is not None:
+            self.configuration_shrink = m.get('Configuration')
         if m.get('Instances') is not None:
             self.instances_shrink = m.get('Instances')
         if m.get('Name') is not None:
@@ -200,7 +200,7 @@ class CreateApplicationShrinkRequest(TeaModel):
         if m.get('TemplateId') is not None:
             self.template_id = m.get('TemplateId')
         if m.get('Variables') is not None:
-            self.variables = m.get('Variables')
+            self.variables_shrink = m.get('Variables')
         return self